refactor(agentboot): split launcher e2e tests into launcher_e2e_test.go
Move 6 real-Claude-CLI tests from launcher_test.go into launcher_e2e_test.go
with //go:build e2e tag, matching the pattern in claude_e2e_test.go.
Removes t.SkipNow() workarounds. Unit tests in launcher_test.go now run fast.
